02-pgvector-private/app/requirements.txt
# requirements.txt - Private RAG Application
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pgvector==0.2.4
sentence-transformers==2.3.1
ollama==0.1.6
python-multipart==0.0.6
pydantic==2.6.0
numpy==1.26.3
